requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-WTF==1.2.1
WTForms==3.1.1
boto3==1.34.7
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
pillow==10.1.0
python-magic==0.4.27
gunicorn==21.2.0
flask-cors==4.0.0
botocore==1.34.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
six==1.16.0
urllib3==2.0.7
b2sdk==1.23.0
wasabi==1.1.2
google-cloud-storage==2.14.0
google-auth==2.27.0
argon2-cffi==23.1.0
cryptography==41.0.7
PyJWT==2.8.0
python-jose==3.3.0